+++ themes-default/slim/src/components/helpers/asset.vue
@@ -53,6 +53,7 @@ export default {
             if (this.link) {
                 return this.src.replace('Thumb', '');
             }
+            return undefined;
         }
     }
 };